# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "imperia-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Mapped server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Current server"
        },
        {
            "url": "https://imperia-api.herokuapp.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Index orders.",
                "operationId": "indexOrders",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/OrderIncludes"
                        }
                    },
                    {
                        "name": "page[size]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageSize"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageNumber"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index orders response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Store order.",
                "operationId": "storeOrder",
                "requestBody": {
                    "description": "Store order request object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create order response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Show orders by id.",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/OrderIncludes"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show order response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "summary": "Delete order.",
                "operationId": "destroyOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Delete request object.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delete order response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DestroyOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "orders"
                ],
                "summary": "Update order.",
                "operationId": "updateOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Update order request object.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Update order response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/banquets/{id}/order": {
            "get": {
                "tags": [
                    "orders",
                    "banquets"
                ],
                "summary": "Show order by banquet id.",
                "operationId": "showOrderByBanquetId",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/OrderIncludes"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the banquet.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show order by banquet id response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/restore": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Restore order.",
                "operationId": "restoreOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the order.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore order response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestoreOrderResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Unauthenticated response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "OrderIncludes": {
                "description": "Coma-separated list of inluded relations. Available relations:\n    `comments`, `spaces.comments`, `tickets.comments`, `products.comments`, `services.comments`",
                "type": "string",
                "example": "comments,spaces.comments,tickets.comments,products.comments,services.comments"
            },
            "IndexOrderResponse": {
                "description": "Index orders response object.",
                "required": [
                    "data",
                    "meta",
                    "message"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ShowOrderResponse": {
                "description": "Show order response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "StoreOrderResponse": {
                "description": "Store order response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "message": {
                        "type": "string",
                        "example": "Created"
                    }
                },
                "type": "object"
            },
            "UpdateOrderResponse": {
                "description": "Update order response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "DestroyOrderResponse": {
                "description": "Delete order response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Deleted"
                    }
                },
                "type": "object"
            },
            "RestoreOrderResponse": {
                "description": "Restore order response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
...
            "PageSize": {
                "description": "Max number of results on page.",
                "type": "integer",
                "example": 25
            },
            "PageNumber": {
                "description": "Page number to be returned.",
                "type": "integer",
                "example": 1
            },
...
            "DestroyRequest": {
                "description": "Destroy request",
                "properties": {
                    "force": {
                        "description": "Determines whether resource should be deleted without chance of restoring.",
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            },
...
            "StoreOrderRequest": {
                "description": "Store order request",
                "required": [
                    "banquet_id"
                ],
                "properties": {
                    "banquet_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "spaces": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestSpaceField"
                        }
                    },
                    "tickets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestTicketField"
                        }
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestServiceField"
                        }
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestProductField"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttachingComment"
                        }
                    },
                    "discounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttachingDiscount"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateOrderRequest": {
                "description": "Update order request",
                "properties": {
                    "spaces": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestSpaceField"
                        }
                    },
                    "tickets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestTicketField"
                        }
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestServiceField"
                        }
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreOrderRequestProductField"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttachingComment"
                        }
                    },
                    "discounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttachingDiscount"
                        }
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "description": "Pagination meta object.",
                "required": [
                    "from",
                    "to",
                    "total",
                    "path",
                    "per_page",
                    "current_page",
                    "last_page",
                    "links"
                ],
                "properties": {
                    "from": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 30
                    },
                    "path": {
                        "nullable": "true",
                        "type": "integer",
                        "example": null
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}